pyproject.toml
[project]
name = "prog-seg-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "cryptography>=43.0.0",
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
    "pyjwt>=2.10.1",
    "bcrypt>=4.0.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "oso>=0.27.3",
    "valkey>=6.1.0",
    "python-jose[cryptography]>=3.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
